requirements.txt
click~=8.1.3
python-can~=4.0.0
bitarray~=2.5.1
inquirer~=2.9.2